refactor: reorganize tracker modules and enhance public API structure
- Renamed and reorganized tracker-related files from `src/tracker/` to `src/trackers/`, improving clarity and consistency in module structure.
- Updated public API exports to reflect the new module paths, including subpath exports for `trackers` in package.json.
- Revised documentation in AGENTS.md, CHANGELOG.md, and README.md to clarify the new structure and usage of tracker modules.
- Enhanced tests to validate the functionality of the updated tracker components and their integration within the analysis pipeline.
