+++ AGENTS.md
@@ -6,14 +6,14 @@ This document orients coding agents working on **chessalyzer.js** — a TypeScri
 
 Chessalyzer.js parses large PGN databases and runs user-defined **trackers** over each game (move-level or game-level statistics). It is designed for throughput on multi-core machines.
 
-**Main entry points:** `parsePGN(path, options?)` in [`src/pgn/parse-pgn.ts`](src/pgn/parse-pgn.ts) (stage 2 only); `analyzePGN(path, options?)` in [`src/core/analyze.ts`](src/core/analyze.ts) (full pipeline).
+**Main entry points:** `analyzePGN(path, options?)` in [`src/core/analyze.ts`](src/core/analyze.ts) (full pipeline); `parsePGN(path, options?)` in [`src/pgn/parse-pgn.ts`](src/pgn/parse-pgn.ts) via `chessalyzer.js/pgn`; trackers via `chessalyzer.js/trackers`.
 
 **Pipeline (high level):**
 
 1. **I/O** — `readLines` / `openLineStream` in [`src/io/line-reader.ts`](src/io/line-reader.ts) and `readPgnChunks` in [`src/io/pgn-chunks.ts`](src/io/pgn-chunks.ts) stream the file with minimal overhead. In multithreaded mode, chunking splits the PGN into byte-sized batches aligned to complete games for worker dispatch (parallel I/O, not a semantic stage).
 2. **PGN parse** — structural parse: tag pairs, mainline SAN strings, game boundaries (`GameAssembler`, [`movetext.ts`](src/pgn/movetext.ts)).
 3. **Replay** — SAN decode + play on a board ([`src/replay/`](src/replay/)), mode `'skip' | 'board' | 'actions'`.
-4. **Analyze** — [`GameProcessor`](src/core/game-processor.ts) runs configured trackers ([`src/tracker/`](src/tracker/)).
+4. **Analyze** — [`GameProcessor`](src/core/game-processor.ts) runs configured trackers ([`src/trackers/`](src/trackers/)).
 
 **Terminology:** Canonical glossary in the [README Pipeline section](README.md#pipeline) and [Sprint 11](sprints/sprint-11-pipeline-terminology.md). Public docs use **replay** for SAN decode + play.
 
@@ -26,7 +26,7 @@ Chessalyzer.js parses large PGN databases and runs user-defined **trackers** ove
 | `src/pgn/`           | PGN parse (`GameAssembler`, `movetext`, `parsePGN`)                                    |
 | `src/replay/`        | Replay — SAN decode + apply (`GameReplayer`, `ReplayMode`, `SanApplier`, `SanDecoder`) |
 | `src/types/`         | Public analysis types (`analysis.ts`) vs processor runtime (`analysis-runtime.ts`)     |
-| `src/tracker/`       | Built-in and base tracker implementations                                              |
+| `src/trackers/`      | Built-in and base tracker implementations                                              |
 | `bench/`             | Callable performance benchmarks (`bench-*.ts`)                                         |
 | `bench/atomic/`      | Atomic micro-benchmark implementations                                                 |
 | `bench/lib/`         | Shared bench utilities (fixtures, timing, PGN resolution)                              |
