+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Export `MoveTracker`, `GameTrackerBase`, and public config/result types.
 - Built-in tracker stats aligned with `AnalyzeResult`: `GameTracker.cntGames` → `games`; `TileTracker.cntMovesGame` / `cntMovesTotal` → `movesGame` / `movesTotal`.
 - Tracker modules renamed: `game-tracker.ts`, `piece-tracker.ts`, `tile/tile-tracker.ts` (drop misleading `-base` suffix on concrete exports).
+- Public API split by subpath: root is analyze-only; `parsePGN` → `chessalyzer.js/pgn`; trackers → `chessalyzer.js/trackers` (named exports).
 
 #### Pipeline terminology (Sprint 11)
 
@@ -46,7 +47,10 @@ Docs and benchmarks use pipeline stage names: **I/O → PGN parse → replay →
 
 - `src/io/` — `line-reader.ts`, `pgn-chunks.ts` (moved from `src/pgn/`)
 - `src/pgn/` — parse-only (`game-assembler`, `movetext`, `parse-pgn`)
-- Package subpath exports: `chessalyzer.js/io`, `chessalyzer.js/pgn`, `chessalyzer.js/replay`
+- Package subpath exports: `chessalyzer.js/io`, `chessalyzer.js/pgn`, `chessalyzer.js/replay`, `chessalyzer.js/trackers`, `chessalyzer.js/trackers`
+- Root export is analyze-only (`analyzePGN`, `printHeatmap`, error helpers/types); `parsePGN` lives on `/pgn`; trackers and tracker types on `/trackers`
+- Built-in trackers use named exports (`GameTracker`, `PieceTracker`, `TileTracker`, `BaseTracker`)
+- Internal folder renamed `src/tracker/` → `src/trackers/` (`#trackers/*` import alias)
 
 **Deferred:**
