+++ src/index.ts
@@ -1,11 +1,4 @@
 export { analyzePGN, printHeatmap } from '#core/analyze';
-export { parsePGN } from '#pgn/parse-pgn';
-export { GameTrackerBase, MoveTracker } from '#tracker/base-tracker';
-export { default as BaseTracker } from '#tracker/base-tracker';
-export { default as PieceTracker, isTrackedPiece } from '#tracker/piece-tracker';
-export { default as GameTracker } from '#tracker/game-tracker';
-export { default as TileTracker } from '#tracker/tile/tile-tracker';
-export type { Action } from '#types/actions';
 export type {
     AnalyzeOptions,
     AnalyzeResult,
@@ -13,10 +6,6 @@ export type {
     AnalyzeRunResult,
     WorkerOptions,
 } from '#types/analysis';
-export type { Game } from '#types/game';
-export type { ParsedGame, ParsePgnOptions } from '#types/parse-pgn';
-export type { HeatmapData, Tracker } from '#types/tracker';
-export type { ReplayMode } from '#replay/replay-policy';
 export type {
     AnalyzeError,
     AnalyzeErrorCode,
