+++ src/trackers/heatmap-utils.ts
