+++ test/integration/error-policy.test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect } from 'bun:test';
 import { join } from 'node:path';
 
-import { analyzePGN, getAnalyzeError, isReplayError, PieceTracker } from 'chessalyzer.js';
+import { analyzePGN, getAnalyzeError, isReplayError } from 'chessalyzer.js';
+import { PieceTracker } from 'chessalyzer.js/trackers';
 
 const badSanPath = join(import.meta.dirname, '../fixtures/bad-san-mid-file.pgn');
