+++ test/integration/workers.test.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect, afterEach } from 'bun:test';
 import { join } from 'node:path';
 
-import { analyzePGN, PieceTracker, TileTracker } from 'chessalyzer.js';
+import { analyzePGN } from 'chessalyzer.js';
+import { PieceTracker, TileTracker } from 'chessalyzer.js/trackers';
 
 import WorkerPool from '../../src/core/worker-pool';
 import { fixturePath } from '../helpers/fixtures';
